feat: move new idea listener to app and clean up
